Remove unused endpoints and amend tests
