+++ tests/test_transform.py
@@ -1,17 +1,11 @@
 import csv
-from datetime import datetime
 import glob
 import io
 import json
 import os
 import unittest
-from unittest.mock import patch
-import zipfile
-
-import dateutil
 
 from transform import app
-from sdx.common.transformer import ImageTransformer
 from transform.views.image_filters import format_date
 
 
@@ -121,60 +115,12 @@ class TestTransformService(unittest.TestCase):
 
             self.assertEqual(actual_response, expected_response)
 
-    @patch('transform.transformers.ImageTransformer.get_image_sequence_numbers', return_value=[1, 2])
-    def test_transforms_csv(self, mock_sequence_no):
-        test_scenarios = get_test_scenarios('csv')
-
-        print("Found %d csv scenarios" % len(test_scenarios))
-
-        for scenario_filename in test_scenarios:
-
-            print("Loading scenario %s " % scenario_filename)
-
-            payload = get_file_as_string(scenario_filename)
-            payload_object = json.loads(payload)
-
-            r = self.app.post(self.transform_images_endpoint, data=payload)
-
-            zip_contents = io.BytesIO(r.data)
-
-            z = zipfile.ZipFile(zip_contents)
-
-            expected_content = get_expected_output(scenario_filename, 'csv')
-            expected_csv = list(csv.reader(io.StringIO(expected_content)))
-
-            date_object = datetime.strptime(expected_csv[0][0], '%d/%m/%Y %H:%M:%S')
-
-            sub_date = dateutil.parser.parse(payload_object['submitted_at'])
-            sub_date_str = sub_date.strftime("%Y%m%d")
-
-            filename = 'EDC_{}_{}_1000.csv'.format(payload_object['survey_id'], sub_date_str)
-
-            self.assertTrue(filename in z.namelist())
-
-            edc_file = z.open(filename)
-
-            actual_content = edc_file.read().decode('utf-8')
-
-            modified_content = modify_csv_time(actual_content, date_object)
-            modified_csv = list(csv.reader(io.StringIO(modified_content)))
-
-            self.assertEqual(expected_csv, modified_csv)
-
     def test_invalid_input(self):
         r = self.app.post(self.transform_pck_endpoint, data="rubbish")
 
         self.assertEqual(r.status_code, 400)
 
         r = self.app.post(self.transform_idbr_endpoint, data="rubbish")
-
-        self.assertEqual(r.status_code, 400)
-
-        r = self.app.post(self.transform_images_endpoint, data="rubbish")
-
-        self.assertEqual(r.status_code, 400)
-
-        r = self.app.post(self.transform_pdf_endpoint, data="rubbish")
 
         self.assertEqual(r.status_code, 400)
 
@@ -188,12 +134,3 @@ class TestTransformService(unittest.TestCase):
         r = self.app.post(self.transform_pck_endpoint, data=payload)
 
         self.assertEqual(r.status_code, 400)
-
-    def test_cleanup(self):
-        for dirpath, dirnames, files in os.walk('./tmp'):
-            if dirnames:
-                ImageTransformer.cleanup(self, './tmp')
-            if not dirnames:
-                self.assertEqual(dirpath, './tmp')
-                self.assertEqual(dirnames, [])
-                self.assertEqual(files, [])
